gamelens/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.8.1
plotly>=5.18.0
wordcloud>=1.9.3
openai>=1.12.0
langdetect>=1.0.9
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.1
